update vendor lib; fix more sensible naming
