+++ colorandcolor.sketchplugin
@@ -1,5 +1,5 @@
-#import 'lib/util.js'
+@import 'lib/gradient.js'
 
-var url = "http://colorandcolor.tumblr.com/random";
+var url = 'http://colorandcolor.tumblr.com/random';
 
-setFillWithUrl(url);
+setFillFromUrl(url);
